move the intent handle logic to import media screen
