+++ app/src/main/kotlin/com/wire/android/util/extension/Context.kt
@@ -1,7 +1,9 @@
 package com.wire.android.util.extension
 
 import android.content.Context
+import android.content.ContextWrapper
 import android.content.pm.PackageManager
+import androidx.appcompat.app.AppCompatActivity
 import androidx.core.content.ContextCompat
 import com.google.android.gms.common.ConnectionResult
 import com.google.android.gms.common.GoogleApiAvailability
@@ -14,4 +16,10 @@ fun Context.checkPermission(permission: String): Boolean {
 fun Context.isGoogleServicesAvailable(): Boolean {
     val status = GoogleApiAvailability.getInstance().isGooglePlayServicesAvailable(this)
     return status == ConnectionResult.SUCCESS
+}
+
+fun Context.getActivity(): AppCompatActivity? = when (this) {
+    is AppCompatActivity -> this
+    is ContextWrapper -> baseContext.getActivity()
+    else -> null
 }
